Added bold and italic values to all message constructors.
Commit for #31 and #43. This commit puts boolean values for bold and italic in all constructors for messages. While it changes a lot of files, it is an atomic commit as splitting up this commit into each individual file change would result in all commits but the final one being unrunnable (attempting to run the application with only some of the files changes but not all will yield build errors).