Implement Delete Conversation Feature
Commit for #62 and #76. When the user confirms their desire to delete the conversation, the messages are removed. Note that the conversation must be exited and reentered for the changes to be seen. Also note that at present, deletes messages show up with message body "...". This will be adressed and fixed in a future commit.