Fixing the pinMessage method
Commit for #32 and #78. The PinMessage and UnpinMessage methods should now be working properly. At this time there is some debug information that is being sent to the console to indicate the state of the pinned variable at various parts of the code. This will be removed prior to merging. All debug information starts with a single letter repeated multiple time, e.g. QQQQQQQ or AAAAAAA.