Make pinned messages work locally
Commit for #32 and #78. The togglePin method takes as input a conversationitem object and toggles the its pinned state (boolean). The pinned state can then be obtained through the isPinned() method. A small filtering system was implemented (and then removed) to test that the method worked. The filtering system was removed because it was only for testing purposes and did not hold any value towards the ultimate completion of the task.