#68 & #69 Fixed Broken Test
Tests were broken due to 1. Firebase configuration was added that was not reflected in unit tests #68, 2. The setup activity had been modified for #68 and the unit tests expected a certain order of execution. Both tests were modified to resolve this issue and a missing listener was added to address #68 problems as well.