Implement a UI for pinning messages
Commit for #77 and #32. Scraps previous idea of having pinned message checkbox on the message objects themselves in the conversation.  Now there is a cyan checkbox in the text editor, if a message is sent while the checkbox is active, the message is sent as pinned. The prior code for checkboxes in the message items were commented out and not removed because we may try again to have the checkboxes on the message items.